rename `serve/` to `internal/`
- import correct package
- move files from `serve/` to `internal/serve`
